Move model code to its dedicated module in root.
